feat: extract engine version logic into custom hook useEngineVersions
